server/requirements.txt
sqlalchemy
pydantic
starlette
fastapi
uvicorn
passlib
bcrypt
google-auth-oauthlib
google-api-python-client
python-dotenv
html2text
ftfy
email-reply-parser
pandas
scikit-learn
transformers